move consent state to common
