+++ apps/docs/features/telemetry/telemetry.client.tsx
@@ -1,11 +1,11 @@
 'use client'
 
 import { IS_PLATFORM, PageTelemetry as PageTelemetryImpl } from 'common'
-import { useConsent } from 'ui-patterns/consent'
+import { useConsentToast } from 'ui-patterns/consent'
 import { API_URL } from '~/lib/constants'
 
 const PageTelemetry = () => {
-  const { hasAcceptedConsent } = useConsent()
+  const { hasAcceptedConsent } = useConsentToast()
 
   return (
     <PageTelemetryImpl
