+++ apps/studio/data/telemetry/send-groups-identify-mutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, UseMutationOptions } from '@tanstack/react-query'
 
 import { components } from 'api-types'
-import { LOCAL_STORAGE_KEYS } from 'common'
+import { hasConsented } from 'common'
 import { handleError, post } from 'data/fetchers'
 import { IS_PLATFORM } from 'lib/constants'
 import type { ResponseError } from 'types'
@@ -9,10 +9,7 @@ import type { ResponseError } from 'types'
 export type SendGroupsIdentifyVariables = components['schemas']['TelemetryGroupsIdentityBody']
 
 export async function sendGroupsIdentify({ body }: { body: SendGroupsIdentifyVariables }) {
-  const consent =
-    (typeof window !== 'undefined'
-      ? localStorage.getItem(LOCAL_STORAGE_KEYS.TELEMETRY_CONSENT)
-      : null) === 'true'
+  const consent = hasConsented()
 
   if (!consent || !IS_PLATFORM) return undefined
