+++ apps/studio/data/telemetry/send-groups-reset-mutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, UseMutationOptions } from '@tanstack/react-query'
 
 import { components } from 'api-types'
-import { LOCAL_STORAGE_KEYS } from 'common'
+import { hasConsented } from 'common'
 import { handleError, post } from 'data/fetchers'
 import { IS_PLATFORM } from 'lib/constants'
 import type { ResponseError } from 'types'
@@ -9,10 +9,7 @@ import type { ResponseError } from 'types'
 export type SendGroupsResetVariables = components['schemas']['TelemetryGroupsResetBody']
 
 export async function sendGroupsReset({ body }: { body: SendGroupsResetVariables }) {
-  const consent =
-    (typeof window !== 'undefined'
-      ? localStorage.getItem(LOCAL_STORAGE_KEYS.TELEMETRY_CONSENT)
-      : null) === 'true'
+  const consent = hasConsented()
 
   if (!consent || !IS_PLATFORM) return undefined
