+++ apps/studio/package.json
@@ -125,7 +125,7 @@
     "ui-patterns": "workspace:*",
     "use-debounce": "^7.0.1",
     "uuid": "^9.0.1",
-    "valtio": "^1.12.0",
+    "valtio": "catalog:",
     "yup": "^1.4.0",
     "yup-password": "^0.3.0",
     "zod": "^3.22.4",
