+++ packages/common/index.tsx
@@ -1,4 +1,5 @@
 export * from './auth'
+export * from './consent-state'
 export * from './constants'
 export * from './database-types'
 export * from './gotrue'
