+++ packages/common/package.json
@@ -12,13 +12,15 @@
   },
   "dependencies": {
     "@types/dat.gui": "^0.7.12",
+    "@usercentrics/cmp-browser-sdk": "^4.42.0",
     "@vercel/flags": "^2.6.0",
     "api-types": "workspace:*",
     "config": "workspace:*",
     "dat.gui": "^0.7.9",
     "lodash": "^4.17.21",
     "next-themes": "^0.3.0",
-    "react-use": "^17.4.0"
+    "react-use": "^17.4.0",
+    "valtio": "catalog:"
   },
   "devDependencies": {
     "@types/lodash": "4.17.5",
