+++ packages/ui-patterns/PrivacySettings/index.tsx
@@ -1,10 +1,9 @@
 'use client'
 
-import { handleResetTelemetry, LOCAL_STORAGE_KEYS } from 'common'
+import { useConsentState } from 'common'
 import Link from 'next/link'
-import { PropsWithChildren, useEffect, useMemo, useState } from 'react'
+import { PropsWithChildren, useState } from 'react'
 import { Modal, Toggle } from 'ui'
-import { useConsentState } from '../consent'
 
 interface PrivacySettingsProps {
   className?: string
