+++ pnpm-workspace.yaml
@@ -13,3 +13,4 @@ catalog:
   'react-dom': '^18.3.0'
   '@types/react': '^18.3.0'
   '@types/react-dom': '^18.3.0'
+  'valtio': '^1.12.0'
