+++ turbo.json
@@ -49,6 +49,7 @@
         "NEXT_PUBLIC_AUTH_NAVIGATOR_LOCK_KEY",
         "NEXT_PUBLIC_AUTH_DETECT_SESSION_IN_URL",
         "NEXT_PUBLIC_VERCEL_ENV",
+        "NEXT_PUBLIC_USERCENTRICS_RULESET_ID",
         // These envs are technically passthrough env vars because they're only used on the server side of Nextjs
         "PLATFORM_PG_META_URL",
         "STUDIO_PG_META_URL",
@@ -105,6 +106,7 @@
         "NEXT_PUBLIC_BASE_PATH",
         "NEXT_PUBLIC_EXAMPLES_SUPABASE_ANON_KEY",
         "NEXT_PUBLIC_EXAMPLES_SUPABASE_URL",
+        "NEXT_PUBLIC_USERCENTRICS_RULESET_ID",
         // These envs are technically passthrough env vars because they're only used on the server side of Nextjs
         "LIVE_SUPABASE_COM_SERVICE_ROLE_KEY",
         "GITHUB_CHANGELOG_APP_ID",
@@ -137,7 +139,6 @@
         "NEXT_PUBLIC_API_URL",
         "NODE_ENV",
         "HIDE_TAILWIND_INDICATOR",
-        "NEXT_PUBLIC_USERCENTRICS_RULESET_ID",
         // These envs are used in the packages
         "NEXT_PUBLIC_STORAGE_KEY",
         "NEXT_PUBLIC_AUTH_DEBUG_KEY",
@@ -145,7 +146,8 @@
         "NEXT_PUBLIC_AUTH_NAVIGATOR_LOCK_KEY",
         "NEXT_PUBLIC_AUTH_DETECT_SESSION_IN_URL",
         "NEXT_PUBLIC_GOTRUE_URL",
-        "NEXT_PUBLIC_SUPABASE_ANON_KEY"
+        "NEXT_PUBLIC_SUPABASE_ANON_KEY",
+        "NEXT_PUBLIC_USERCENTRICS_RULESET_ID"
       ],
       "outputs": [".next/**", "!.next/cache/**", ".contentlayer/**"]
     },
